Dev -> Prod (#164)
* abn - number closes #139

* closes #146

* Remove Settings - closes #151

* Footer - closes #149

* closes #150

* Nsg223/pre launch fixes (#158)

* closes #148

* misc changes around search - working on #128

* Billing Error handling and logging

* fix #152

* resolve #153

* fix #147

* fix #157

* fix comments button for admins

* fix seed data

* fix #145

* Nsg223/pre launch fixes (#159)

* closes #148

* misc changes around search - working on #128

* Billing Error handling and logging

* fix #152

* resolve #153

* fix #147

* fix #157

* fix comments button for admins

* fix seed data

* fix #145

* fix seed info for demail confirrmed

* Pre Launch Fixes (#162)

* fix profile not showing un accepted job offers

* more fixes

* minor home page fix

* Closes #160 Removed Email Redirect

* Closed multiple issues; see description for notes

Delete button removed from employer view of employees; closes #155
Job offer controller security updated; closes #161
E-mail text updated; closes #138
Activation proccess completed; closes #114

* Finalisation liam (#163)

* pricing page / email updates

Added to pricing page - will be linked from marketing site; closes #140
Added radio buttons in this branch on creating job offers;
Fixed e-mail for employers (welcome);
Changed view accepted offers button on employer page;

* Restored data preservation on joboffers

* secondary job field properly returns null unselected

* Secured controllers

Also restored admin functionality of deleting comments

* Rendancy removed

+++ app/Http/Controllers/JobofferController.php
@@ -49,29 +49,38 @@ class JobofferController extends Controller
     }
 
     public function destroy($id)
-    {
+    {   
         $joboffer = Joboffer::find($id);
+        if(Auth::user()->employer->id == $joboffer->employer_id){
         $joboffer->delete();
-        return redirect('/');
+        return back()->with("success", "Job Offer Deleted");
+        }
+        else
+        return ('home');
     }
 
     public function create()
-    {
+    {    
         $roles = Settings::$roles;
         return view('joboffers.create', compact('roles'));
     }
 
     public function edit($id){
         $joboffer = Joboffer::find($id);
+        if(Auth::user()->employer->id == $joboffer->employer_id){
         //want mm/dd/yyyy, have , dd/mm/yyyy
         $dates = explode('-', $joboffer->date);
         $date = $dates[2].'/'.$dates[1].'/'.$dates[0];
         return view('joboffers.edit', compact('joboffer', 'date'));
+        }
+        else
+        return ('home');
     }
 
     public function update(Request $request, $id)
     { 
         $joboffer = Joboffer::find($id);
+        if(Auth::user()->employer->id == $joboffer->employer_id){
         $dates = explode('/', $request->date);
         $realdate = $dates[2].'-'.$dates[1].'-'.$dates[0];  
         $this->validate($request, Joboffer::$rules);
@@ -81,6 +90,8 @@ class JobofferController extends Controller
         $joboffer->date = $realdate;
         $joboffer->save();
         return redirect('/jobs');
+        }
+        else return('home');
     }
 
 }
